history_sync_patch: fix the search box resetting
After a few searches. After reaching the maximum history count (about 50 by default), the combo box would remove the searched item and replace it with the last entry in the history. The fix is to add the last searched item to the history, set the item in the combo box, and reset the history to sync.

reset() is required according to the KDE API Doc. http://api.kde.org/frameworks-api/frameworks5-apidocs/kcompletion/html/classKHistoryComboBox.html#af6ef37548be290c810e1603350393780