Call KLocalizedString::setApplicationDomain("kiten");
kiten is using
   -DTRANSLATION_DOMAIN=\"kiten\"
that is wrong for applications, apps need to use
   KLocalizedString::setApplicationDomain
since that is needed for some things like the transltators tab in the about
dialog.

I have not removed -DTRANSLATION_DOMAIN=\"kiten\" because on the other hand
this code is also a library.

The "being pedanticly correct" fix would be using -DTRANSLATION_DOMAIN=\"kiten\"
only to compile the files that are part of the library but that's prone to
break at some point and using both KLocalizedString::setApplicationDomain and
-DTRANSLATION_DOMAIN does not create any problem so i went that route