Port to QTextBrowser
Summary:
This ports away from the KHTML dependency. Note that QTextBrowser
does not support CSS pseudo class selectors, which means that the
hover property cannot be set for CSS elements. Thus the results view
is changed such that the light-gray background is not bound to the
hovering of a mouse courser but is applied to every odd line.

Reviewers: yurchor

Reviewed By: yurchor

Subscribers: yurchor, kde-edu

Tags: #kde_edu

Maniphest Tasks: T11574

Differential Revision: https://phabricator.kde.org/D25437